+++ packages/core/src/codewhisperer/commands/startTransformByQ.ts
@@ -27,7 +27,6 @@ import {
     downloadHilResultArchive,
     findDownloadArtifactStep,
     getArtifactsFromProgressUpdate,
-    getTransformationPlan,
     getTransformationSteps,
     pollTransformationJob,
     resumeTransformationJob,
@@ -554,28 +553,6 @@ export async function pollTransformationStatusUntilPlanReady(jobId: string, prof
         // for now, no plan shown with SQL conversions. later, we may add one
         return
     }
-    let plan = undefined
-    try {
-        plan = await getTransformationPlan(jobId, profile)
-    } catch (error) {
-        // means API call failed
-        getLogger().error(`CodeTransformation: ${CodeWhispererConstants.failedToCompleteJobNotification}`, error)
-        transformByQState.setJobFailureErrorNotification(
-            `${CodeWhispererConstants.failedToGetPlanNotification} ${(error as Error).message}`
-        )
-        transformByQState.setJobFailureErrorChatMessage(
-            `${CodeWhispererConstants.failedToGetPlanChatMessage} ${(error as Error).message}`
-        )
-        throw new Error('Get plan failed')
-    }
-
-    if (plan !== undefined) {
-        const planFilePath = path.join(transformByQState.getProjectPath(), 'transformation-plan.md')
-        fs.writeFileSync(planFilePath, plan)
-        await vscode.commands.executeCommand('markdown.showPreview', vscode.Uri.file(planFilePath))
-        transformByQState.setPlanFilePath(planFilePath)
-        await setContext('gumby.isPlanAvailable', true)
-    }
     jobPlanProgress['generatePlan'] = StepProgress.Succeeded
     throwIfCancelled()
 }
@@ -699,11 +676,14 @@ export async function postTransformationJob() {
     }
 
     let chatMessage = transformByQState.getJobFailureErrorChatMessage()
-    const diffMessage = CodeWhispererConstants.diffMessage(transformByQState.getMultipleDiffs())
     if (transformByQState.isSucceeded()) {
-        chatMessage = CodeWhispererConstants.jobCompletedChatMessage(diffMessage)
+        chatMessage = CodeWhispererConstants.jobCompletedChatMessage
     } else if (transformByQState.isPartiallySucceeded()) {
-        chatMessage = CodeWhispererConstants.jobPartiallyCompletedChatMessage(diffMessage)
+        chatMessage = CodeWhispererConstants.jobPartiallyCompletedChatMessage
+    }
+
+    if (transformByQState.getSourceJDKVersion() !== transformByQState.getTargetJDKVersion()) {
+        chatMessage += CodeWhispererConstants.upgradeLibrariesMessage
     }
 
     transformByQState.getChatControllers()?.transformationFinished.fire({
@@ -731,16 +711,23 @@ export async function postTransformationJob() {
         })
     }
 
+    let notificationMessage = ''
+
     if (transformByQState.isSucceeded()) {
-        void vscode.window.showInformationMessage(CodeWhispererConstants.jobCompletedNotification(diffMessage), {
+        notificationMessage = CodeWhispererConstants.jobCompletedNotification
+        if (transformByQState.getSourceJDKVersion() !== transformByQState.getTargetJDKVersion()) {
+            notificationMessage += CodeWhispererConstants.upgradeLibrariesMessage
+        }
+        void vscode.window.showInformationMessage(notificationMessage, {
             title: localizedText.ok,
         })
     } else if (transformByQState.isPartiallySucceeded()) {
+        notificationMessage = CodeWhispererConstants.jobPartiallyCompletedNotification
+        if (transformByQState.getSourceJDKVersion() !== transformByQState.getTargetJDKVersion()) {
+            notificationMessage += CodeWhispererConstants.upgradeLibrariesMessage
+        }
         void vscode.window
-            .showInformationMessage(
-                CodeWhispererConstants.jobPartiallyCompletedNotification(diffMessage),
-                CodeWhispererConstants.amazonQFeedbackText
-            )
+            .showInformationMessage(notificationMessage, CodeWhispererConstants.amazonQFeedbackText)
             .then((choice) => {
                 if (choice === CodeWhispererConstants.amazonQFeedbackText) {
                     void submitFeedback(